A bunch of cleanup.
M    src/CGtest.hpp
M    src/SymTest.cpp
A    src/NormTest.cpp
M    src/SymTest.hpp
A    src/NormTest.hpp
M    src/YAML_Doc.cpp
M    src/CG.cpp
M    src/ReportResults.cpp
M    src/Geometry.hpp
M    src/ReportResults.hpp
M    src/CGref.cpp
M    src/CGtest.cpp
M    testing/main.cpp


git-svn-id: svn+ssh://software.sandia.gov/svn/hpcg/trunk@120 5e06a57e-6b39-4354-92a0-249a7d0c5410

+++ src/Geometry.hpp
@@ -11,8 +11,8 @@
 #ifndef GEOMETRY_HPP
 #define GEOMETRY_HPP
 
-typedef int local_int_t;      // This defines the type for integers that have local subdomain dimension
-//typedef long long local_int_t;      // Define as long long when local problem dimension is > 2^31
+//typedef int local_int_t;      // This defines the type for integers that have local subdomain dimension
+typedef long long local_int_t;      // Define as long long when local problem dimension is > 2^31
 
 // This macro should be defined if the global_int_t is not long long
 // in order to stop complaints from non-C++11 compliant compilers
